Allows UI to config plugins' settings ((un)loading plugins to follow)
    Replaces hardcoded FEM_TOTAL_PIXELS variable with JSON file configurable fem_max_rows, fem_max_cols variables determining fem_total_pixels_
    Ensures that variables for filenames are universally called filename, not file
    Ensures variables filled by JSON file have matching variable type
    Fixes threshold mode is properly set and read; threshold parameters following naming convention
    Odin UI rows, columns now updates height and width of all plugins
    Implements UI to odin_server (and onto FP) communications

+++ data/frameProcessor/src/HexitecHistogramPlugin.cpp
@@ -146,7 +146,7 @@ namespace FrameProcessor
 
     if (config.has_param(HexitecHistogramPlugin::CONFIG_BIN_START))
     {
-    	bin_start_ = config.get_param<long>(HexitecHistogramPlugin::CONFIG_BIN_START);
+    	bin_start_ = config.get_param<int>(HexitecHistogramPlugin::CONFIG_BIN_START);
 		}
 
     if (config.has_param(HexitecHistogramPlugin::CONFIG_BIN_END))
